requirements.txt
flask
flask-cors
requests
gunicorn
aiogram
python-dotenv
